pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viralforge-ai"
version = "1.0.0"
description = "Local-first short-form content automation pipeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "moviepy>=1.0.3",
  "ffmpeg-python>=0.2.0",
  "chromadb>=0.5.5",
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.3",
  "playwright>=1.47.0",
  "edge-tts>=6.1.12",
  "pillow>=10.4.0",
  "rich>=13.7.1",
  "flask>=3.0.3",
  "numpy>=1.26.4",
  "requests-oauthlib>=1.3.1",
]

[project.scripts]
viralforge = "viralforge.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = [
  "agents*",
  "analytics*",
  "application*",
  "interfaces*",
  "llm_router*",
  "memory*",
  "monetization*",
  "posting*",
  "research*",
  "self_healing*",
  "subtitles*",
  "utils*",
  "video_engine*",
  "viralforge*",
]
